research-assistant/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6

# OpenAI and AI
openai==1.3.0
tiktoken==0.5.1

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1

# Redis and Caching
redis==5.0.1
aioredis==2.0.1

# Vector Search
faiss-cpu==1.7.4
numpy==1.24.4
scipy==1.11.4

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3

# Web Scraping
aiohttp==3.9.1
beautifulsoup4==4.12.2
lxml==4.9.3

# Data Processing
pandas==2.1.3
networkx==3.2.1

# Configuration
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Logging and Monitoring
structlog==23.2.0
rich==13.7.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.0
pre-commit==3.5.0